+++ src/main/java/controller/Controller.java
@@ -1,7 +1,7 @@
 package controller;
 
-import webserver.HttpRequest;
-import webserver.HttpResponse;
+import http.HttpRequest;
+import http.HttpResponse;
 
 import java.io.IOException;
 import java.net.URISyntaxException;
