+++ src/main/java/controller/StaticController.java
@@ -1,8 +1,8 @@
 package controller;
 
 import utils.FileIoUtils;
-import webserver.HttpRequest;
-import webserver.HttpResponse;
+import http.HttpRequest;
+import http.HttpResponse;
 
 import java.io.IOException;
 import java.net.URISyntaxException;
