+++ src/main/java/http/HttpParameters.java
@@ -1,4 +1,4 @@
-package webserver;
+package http;
 
 import java.util.HashMap;
 import java.util.Map;
